+++ src/Products/MeetingCharleroi/tests/testCustomMeetingItem.py
@@ -95,7 +95,6 @@ class testCustomMeetingItem(MeetingCharleroiTestCase, mctcmi):
         cfg.setItemAutoSentToOtherMCStates(('itemfrozen', ))
         # create 2 College meetings, one extraordinarySession and one normal session
         # then send an item from each to a Council meeting
-        self.changeUser('pmManager')
         # create the Council meeting
         self.setMeetingConfig(cfg2Id)
         self.setupCouncilConfig()
@@ -134,7 +133,6 @@ class testCustomMeetingItem(MeetingCharleroiTestCase, mctcmi):
 
         self.presentItem(itemFromExtraSession)
         self.assertEqual(itemFromExtraSession.getListType(), 'lateextracollege')
-
         self.presentItem(communicationitem)
         self.assertNotEqual(communicationitem.getListType(), 'lateextracollege')
 
@@ -158,7 +156,6 @@ class testCustomMeetingItem(MeetingCharleroiTestCase, mctcmi):
                     title='Communications')
         gic2_uid = org_id_to_uid('groupincharge2')
         dev_group_in_charge = '{0}__groupincharge__{1}'.format(self.developers_uid, gic2_uid)
-
 
         self.changeUser('pmManager')
         councilMeeting = self.create('Meeting', date=DateTime('2017/01/01').asdatetime())
@@ -385,19 +382,36 @@ class testCustomMeetingItem(MeetingCharleroiTestCase, mctcmi):
 
         # as item is to send to Council, category 'indeterminee' must be used
         self.failIf(item.validate_category(COUNCIL_DEFAULT_CATEGORY))
-        self.assertEqual(item.validate_category('another_category'), msg_mandatory)
+        self.assertEqual(item.validate_category('development'), msg_mandatory)
 
         # but can not be used for items not to send to Council
         item.setOtherMeetingConfigsClonableTo(())
         self.assertEqual(item.validate_category(COUNCIL_DEFAULT_CATEGORY), msg_not_allowed)
-        self.failIf(item.validate_category('another_category'))
+        self.failIf(item.validate_category('development'))
 
         # does not fail when used on MeetingItemCouncil
         self.setMeetingConfig(self.meetingConfig2.getId())
+        # make sure the COUNCIL_DEFAULT_CATEGORY exists or it does not pass validation
+        self.changeUser('siteadmin')
+        self._createCategories()
         self.changeUser('pmManager')
         council_item = self.create('MeetingItem')
         self.assertEqual(council_item.portal_type, 'MeetingItemCouncil')
         self.failIf(council_item.validate_category(COUNCIL_DEFAULT_CATEGORY))
+
+    def test_KeepPollTypeIfCollegeItemToSendToCouncil(self):
+        """When item is sent to Council, keep the poll type."""
+        cfg = self.meetingConfig
+        cfg.setItemManualSentToOtherMCStates(('itemcreated', ))
+        cfg2Id = self.meetingConfig2.getId()
+
+        self.changeUser('pmCreator1')
+        item = self.create('MeetingItem')
+        item.setPollType('secret_separated')
+        item.setOtherMeetingConfigsClonableTo((cfg2Id,))
+        councilItem = item.cloneToOtherMeetingConfig(cfg2Id)
+        self.assertEqual(item.getPollType(), 'secret_separated')
+        self.assertEqual(councilItem.getPollType(), 'secret_separated')
 
 
 def test_suite():